Restructure on handleDocumentDelete api call, moved to the parent comp
